fixed postgres getAvailableDaysFromProfessional
