+++ src/main/resources/application-test.properties
@@ -6,3 +6,5 @@ spring.datasource.password=
 
 spring.jpa.defer-datasource-initialization=true
 spring.sql.init.encoding=UTF-8
+
+spring.flyway.enabled=false
